requirements/production.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anthropic==0.64.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.0.1
autoflake==2.3.1
autopep8==2.3.2
bandit==1.8.6
beautifulsoup4==4.13.5
black==25.1.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
coverage==7.10.6
cryptography==45.0.6
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
docopt==0.6.2
einops==0.8.1
feedparser==6.0.11
filelock==3.19.1
flake8==7.3.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hdbscan==0.8.40
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
HyperPyYAML==1.2.2
identify==2.6.13
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.9
-e git+https://github.com/msg43/Knowledge_Chipper.git@c823be0c53c2bea5df72ac56f09b4ed977727f66#egg=knowledge_system
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.4
lightning-utilities==0.15.2
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mypy==1.17.1
mypy_extensions==1.1.0
narwhals==2.2.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
openai==1.102.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
postgrest==1.1.1
pre_commit==4.3.0
primePy==1.3
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyannote-audio==4.0.1
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.3.0
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydocstyle==6.3.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.3.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.4
pytorch-metric-learning==2.9.0
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.7.0
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
semver==3.0.4
sentence-transformers==5.1.0
sentencepiece==0.2.1
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
speechbrain==1.0.3
SQLAlchemy==2.0.43
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
stevedore==5.5.0
storage3==0.12.1
streamlit==1.49.0
streamlit-option-menu==0.4.0
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchcodec==0.8.0
torchmetrics==1.8.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
typer==0.17.3
types-psutil==7.0.0.20250822
types-PyYAML==6.0.12.20250822
types-requests==2.32.4.20250809
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
youtube-transcript-api==1.2.2
yt-dlp==2025.10.14
zipp==3.23.0
